V2 FixedRateIrm.sol
Replaced string constants for error handling with custom error types. This is more gas-efficient and recommended for Solidity 0.8.0 and later.

Adjusted the constant MAX_BORROW_RATE to use scientific notation 8e18 for better readability.

Added the SetBorrowRate event declaration to the contract, which is necessary for the event to be emitted properly.

Replaced require statements with revert statements using the custom errors for better gas efficiency and clarity.

Improved documentation comments to clarify the purpose of the contract and functions.

Signed-off-by: JamesPotterr <161749116+JamesPotterr@users.noreply.github.com>